requirements.txt.txt
cassandra-driver==3.24.0
certifi==2020.4.5.1
chardet==3.0.4
cheroot==8.3.0
CherryPy==18.6.0
click==7.1.2
cycler==0.10.0
Flask==1.1.2
geomet==0.2.1.post1
idna==2.9
itsdangerous==1.1.0
jaraco.classes==3.1.0
jaraco.collections==3.0.0
jaraco.functools==3.0.1
jaraco.text==3.2.0
Jinja2==2.11.2
kiwisolver==1.2.0
linecache2==1.0.0
MarkupSafe==1.1.1
matplotlib==3.2.2
more-itertools==8.3.0
numpy==1.18.4
pandas==1.0.3
Paste==3.4.1
portend==2.6
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
requests==2.23.0
six==1.14.0
tempora==3.0.0
traceback2==1.4.0
urllib3==1.25.9
Werkzeug==1.0.1
zc.lockfile==2.0